requirements.txt
streamlit>=1.25.0
pandas>=2.0.0
requests>=2.28.0
beautifulsoup4>=4.12.2
pillow>=10.0.0
openpyxl>=3.1.0
